// src/client/models/CreateServiceSchema.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CreateServiceSchema = {
    title: string;
    offices: Array<string>;
    description: string;
    masters_count: number;
    duration: number;
    duration_before: number;
    duration_after: number;
    company_id: string;
};