// src/client/models/ServiceSchema.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ServiceSchema = {
    title: string;
    offices: Array<string>;
    description: string;
    masters_count: number;
    duration: number;
    duration_before: number;
    duration_after: number;
    company_id: string;
    id: string;
};